requirements.txt
flask
flask-restful
passlib
requests
flask-cors
gunicorn==20.0.0
facenet-pytorch
opencv-python-headless
Pillow>=8.0.1
pybase64
pytorch-lightning>=1.0.8
torch>=1.7.0
facenet-pytorch
torchvision>=0.8.1
pickle-mixin
scikit-learn
gdown>=3.12.2


